+++ gui/src/data/mod.rs
@@ -3,8 +3,6 @@ pub use album::*;
 
 mod state;
 pub use state::*;
-mod state0;
-pub use state0::*;
 
 mod settings;
 pub use settings::*;