+++ CHANGELOG.md
@@ -17,11 +17,20 @@ While `festival-gui` is the only `Frontend` available, the changelog and release
 
 
 ## Unreleased
+Performance improvements. Expect 2x~ faster Collection resets (user time) and 5x~ faster overall time (including save).
+
 ## Changed
-* Faster `JPG` album art image decoding, 1.75x~ faster Collection reset ([#20](https://github.com/hinto-janai/festival/pull/20))
+* `JPG` album art decoding now uses [`zune-jpeg`](https://github.com/etemesi254/zune-image), 1.75x~ faster `Collection` reset ([#20](https://github.com/hinto-janai/festival/pull/20))
+* `Collection` directories are now pre-emptively cached on startup and addition; initial reset speeds are faster ([#30](https://github.com/hinto-janai/festival/pull/30))
+* Album art conversion now uses all available threads, 1.25x~ faster ([#20](https://github.com/hinto-janai/festival/pull/20))
+* Post-`Collection` reset image encoding/save can now use multiple threads, 5x~ faster ([#37](https://github.com/hinto-janai/festival/pull/37))
+* Search can now use multiple threads, 1.8x~ faster ([#37](https://github.com/hinto-janai/festival/pull/37))
 
 ## Fixed
 * Crashes with songs that have odd date metadata, again (https://github.com/hinto-janai/readable/commit/02bdd467363e50627e68af56497eaeb13cdf632d)
+* Over-saturated colors on Linux (KDE) ([#43](https://github.com/hinto-janai/festival/pull/43))
+* Runtime UI overflow when song is longer than an hour ([#44](https://github.com/hinto-janai/festival/pull/44))
+
 
 ---